+++ tests/Requests/RequestTest.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Dsone\Strainex\Tests\Request;
+
+use Dsone\Strainex\Tests\TestCase;
+
+class RequestTest extends TestCase
+{
+	/**
+	 * A basic test example.
+	 *
+	 * @return void
+	 */
+	public function test_a_basic_request()
+	{
+		$this->assertTrue(true);
+	}
+}