+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # CHANGELOG
 
-## 0.2 - 06.03.2021
+## 0.3 - 2021-05-08
+
+- [ch445] Add option for IP hashing
+- [ch441] Add filter option for user Agents
+- [ch441] Change `$type` parameter in callbacks to `$criteriaBits`
+- Change namespace from `Dsone\ExceptionHandler` to `Dsone\Strainex`
+
+## 0.2 - 2021-03-06
 
 - Supporting map table instead of sequential array for SEO spam referer list
 - Support for `exit` instead of returning response codes
 - Adding PHP 7.3 as dependency
 - Formatting/expanding README
 
-## 0.1 - 06.03.2021
+## 0.1 - 2021-03-06
 
 - Initial version